refactor: move link logic to a separate module
