+++ src/lib.rs
@@ -1,2 +1,5 @@
+pub mod cmd;
 pub mod config;
+pub mod directories;
 pub mod error;
+pub mod linker;
